Fixed skeleton info error (I changed V4 from int to float compartments, but hadn't changed the info to reflect the change -- bug now fixed).

+++ skeletonservice/datasets/service.py
@@ -90,7 +90,7 @@ SKELETON_VERSION_PARAMS = {
             'num_components': 1,
         },{
             'id': 'compartment',
-            'data_type': 'uint8',
+            'data_type': 'float32',
             'num_components': 1,
         }]},
 }
